requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
deepl==1.18.0
et-xmlfile==1.1.0
idna==3.7
keyboard==0.13.5
openpyxl==3.1.5
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
